Update Dockerfile and build.spec for JDK 17 compatibility
This commit makes several changes to ensure the project is compatible with JDK 17. In the Dockerfile, the base images have been updated from JDK 19 to JDK 17 and also for the final image, we've switched from gradle:jdk19 to centos with manual install of JDK 19. The docker-compose.yaml file was also updated to reduce memory limits. In addition, the build.gradle.kts file has been updated to change java sourceCompatibility and jvmTarget from version 19 to version 17. This change was necessary to ensure the project can be compiled and run with JDK 17.