Add package.json and yarn.lock with project dependencies
This commit adds a new package.json file, containing all the metadata relevant to the project and the necessary packages, and a new yarn.lock file to keep track of the exact package version used for each dependency.

The reason for this change is to setup the project with the necessary dependencies to build an express node.js application that uses 'puppeteer' to generate PDF from HTML, 'nodemon' to automatically restarting the node server, and 'ts-node' to execute typescript code. It also included some development dependencies needed for coding and linting in typescript.