Update dependencies, JDK version, and PDF generation logic
This commit updates the project to JDK 19, changes Selenium to Retrofit for PDF generation, and modifies Dockerfiles for new setup. The change in architecture from Selenium to Retrofit-based server helps to reduce resource consumption, optimize performance, and simplify scaling. The implementation("org.seleniumhq.selenium:selenium-java:4.15.0") dependency in build.gradle.kts was removed in favour of additional Retrofit and micronaut-http-client dependencies. Similarly, the ChromeDriver usage in PdfGeneratorController.kt was removed, with the PDF generation logic now conducted through a RetrofitService interface. Docker-compose.yaml, Dockerfile, and Makefile have been modified to suit the new structure/setup. The Dockerfile now includes JDK 19 and Node.js and changes the execution of the project to start the Node.js server before the Java server. The Node.js PDF generation server was added to the docker-compose.yaml file and a new boot bash script was created for running the two servers. Changes were made to related files like package.json, tsconfig.json, and PDFGeneratorService.ts in the node.js project to ensure the usage of TypeScript and the successful removal of an SVG element before generating the PDF.