Remove redundant files and add Dockerfile and Runner
The relevant JavaScript and TypeScript script files and references were removed in this revision.
Also, a Dockerfile along with a runner script were introduced in the changes to facilitate the deployment
of the application within a Docker container. Some modification in the Makefile, .gitignore file was
also done. The aim of this commit is to remove unnecessary files and setup Docker environment for
the cleaner deployment and enhancement in system performance.