Refactor PDF service for better Docker compatibility
Performed necessary refactoring in the PDF generator service to achieve better compatibility when running it inside a Docker container. Introduced a dedicated method for launching the Broswer instance with appropriate headless options. This change mainly occurs in 'PDFGeneratorService.js'. In addition, the application permissions in the Dockerfile were adjusted to ensure smoother execution in the Docker container. Furthermore, minor changes were made to several files ('app.js', 'Router.js', 'PageLoadConfig.js') for cleaning up the code and some files were removed to maintain the project's simplicity. Refactoring was done to improve the maintainability and readability of the code, and to optimize Docker container performance.