Remove boot file, update Dockerfile, and add bootApp script
The boot file was removed and replaced by a new file - the bootApp script. This was done to improve the way the PDF service is started in a Docker container, resulting in better compatibility and performance. The Dockerfile has been updated to reflect these changes in the starting procedure. The main change was moving the initial yarn installation and the start of the node app to the bootApp script file to decouple the build and the run stages.