Update Java version, allocate higher memory and optimize PDF generation
Upgraded source compatibility and jvmTarget for Kotlin compilation from version 17 to 19 in build.gradle.kts. Resources limit in docker-compose.yaml was increased from 254M to 1G, in order to support more intensive tasks. In PdfGeneratorController.kt, ChromeDriver instance was converted to class property from being created and closed within the function for better memory management. Added shell script to overload the server for testing purposes. Dockerfile was updated by changing the JDK to 19 and including a fixed version of chrome for a more reliable build.